refactor(web): extract tour landing components в components/tour/* (10-15 туров готовность)
С учётом 10-15 туров на старте — дробим монолит page.tsx (610 строк) на
переиспользуемые компоненты. Один template рендерит все туры; кастомизация
per-tour будет проще когда design system подвинется.

Структура apps/web/components/tour/:
- _shared.tsx       — SectionHead, formatRub, formatPriceLabel
- Hero.tsx          — full-viewport с LCP image, sticky price, 2 CTA
- Facts.tsx         — 6 quick facts с iconKind→ICON_MAP
- DayTimeline.tsx   — accordion дней + ACTIVITY_ICONS + pluralizeDays
- Inclusions.tsx    — что входит / не входит
- Reviews.tsx       — текстовые отзывы
- PriceBlock.tsx    — CTA + leadFormSlot prop
- FAQ.tsx           — accordion вопросы
- FooterLegal.tsx   — юр.ссылки + ИНН/ОГРН placeholder

Изменения по содержанию:
- DayTimeline title теперь динамический: "10 дней — без хаоса" → tour.durationDays
  через pluralizeDays() (1/2-4/5-20 склонения). Раньше hardcoded "10 дней" —
  было бы багом для туров другой длительности.
- PriceBlock принимает leadFormSlot: ReactNode — dynamic-imported LeadForm
  передаётся слотом со страницы (RSC-friendly).
- FooterLegal: убрал устаревшую пометку "(в работе #307)" — страницы DRAFT уже опубликованы.

page.tsx: 610 → 134 строк. Bundle size без изменений (pure refactor).

Test plan:
- [x] pnpm lint, format:check, type-check, build — green
- [x] Bundle size /tours/[slug]: 8.73 kB / 95.9 kB First Load JS — без регрессии
- [x] Schema.org SEO sigs остались в page.tsx (только оборачивающий <main>)
